requirements.txt
PyQt6>=6.6.0
keyboard>=0.13.5
requests>=2.31.0
watchdog>=4.0.0
pywin32>=306
Pillow>=10.0.0
mss>=9.0.0
